Added CORS Support and EC2 Support
This commit mainly adds support for 2 features. First, it enables CORS support
so that requests can be made across domain names. Second, it adds a script
for deploying the job-server to EC2. The commit also includes an example to
run on the EC2 cluster. Finally, there are improvements in the bin/ scripts
to better support paths with spaces in them. Previously, these paths had
caused issues with the bin/sever_deploy.sh script.